requirements.txt
streamlit>=1.32.0
google-generativeai>=0.7.0
pydantic>=2.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.18.0
python-pptx>=0.6.21
openpyxl>=3.1.0
